+++ utils/renderer.py
@@ -67,7 +67,7 @@ def draw_skeleton(input_image, joints, draw_edges=True, vis=None, radius=None):
     image = input_image.copy()
     input_is_float = False
 
-    if np.issubdtype(image.dtype, np.float):
+    if np.issubdtype(image.dtype, np.float32):
         input_is_float = True
         max_val = image.max()
         if max_val <= 2.:  # should be 1 but sometimes it's slightly above 1
@@ -187,7 +187,7 @@ def draw_text(input_image, content):
     """
     image = input_image.copy()
     input_is_float = False
-    if np.issubdtype(image.dtype, np.float):
+    if np.issubdtype(image.dtype, np.float32):
         input_is_float = True
         image = (image * 255).astype(np.uint8)
